chore: refactored bio card to use the card pieces.
