+++ components/elements/buttons/SocialButtons/SocialButtons.test.ts
@@ -1,5 +1,5 @@
 import SocialButtons from './SocialButtons.vue';
-import { stubbedRender } from '@/testUtils';
+import { stubbedRender } from '~/testUtils';
 
 describe('<SocialButtons />', () => {
   const defaultProps = {
